+++ packages/sources/src/components/navbar/navbar.models.ts
@@ -1,0 +1,11 @@
+export interface Navbar {
+  items: NavbarItem[];
+  modifier: string;
+  open: boolean;
+}
+
+export interface NavbarItem {
+  href: string;
+  label: string;
+  active?: boolean;
+}